seperate db into multiple by functionality and implement a raw function to export log to csv
