+++ db/db.js
@@ -1,724 +1,729 @@
 'use strict';
 var Sequelize = require('sequelize');
 var models = require('../models');
-const utils = require('../utils/logging');
-const uuid = require('uuid/v4');
+// const utils = require('../utils/logging');
+// const uuid = require('uuid/v4');
 const sequelize = models.sequelize;
+
 const getter = require('./db_getter');
+const adder = require('./db_adder');
+const remover = require('./db_remover');
+const updater = require('./db_updater');
 
 sequelize.sync();
-
-const perror = utils.perror;
-const info = utils.info;
-const log = utils.log;
-
-const Op = Sequelize.Op;
-const CourseOffering = models.CourseOffering;
-const Course = models.Course;
-const Dept = models.Dept;
-const EchoSection = models.EchoSection;
-const Lecture = models.Lecture;
-const Media = models.Media;
-const MSTranscriptionTask = models.MSTranscriptionTask;
-const Offering = models.Offering;
-const Role = models.Role;
-const Term = models.Term;
-const University = models.University;
-const User = models.User;
-const UserOffering = models.UserOffering;
-const YoutubeChannel = models.YoutubeChannel;
-const CourseOfferingMedia = models.CourseOfferingMedia;
-/* ----- end of defining ----- */
-
-// function getAllCourses() {
-//   return Course.findAll({limit: 8})
-// }
-
-function addCourseOfferingMedia(courseOfferingId, mediaId, description) {
-    return CourseOfferingMedia.findOrCreate({
-      where: {
-        courseOfferingId: courseOfferingId,
-        mediaId: mediaId
-      },
-      defaults: {
-          id: uuid(),
-          descpJSON: JSON.stringify(description),
-          mediaId: mediaId,
-          courseOfferingId: courseOfferingId
-      }
-    })
-}
-
-// function getPlaylistByCourseOfferingId(courseOfferingId) {
-//   return sequelize.query(
-//    'SELECT mst.videoLocalLocation, mst.srtFileLocation, M.siteSpecificJSON \
-//     FROM MSTranscriptionTasks AS mst \
-//     INNER JOIN Media as M on mst.mediaId = M.id \
-//     INNER JOIN CourseOfferingMedia as com on com.mediaId = M.id \
-//     WHERE com.courseOfferingId = ?',
-//    { replacements: [ courseOfferingId ], type: sequelize.QueryTypes.SELECT}).catch(err => perror(err)); /* raw query */
-// }
-
-function addYoutubeChannelPlaylist(playlistId, channelId) {
-    return YoutubeChannel.findOrCreate({
-        where: {
-            playlistId: playlistId
-        },
-        defaults: {
-            playlistId: playlistId,
-            channelId: channelId
-        }
-    });
-}
-
-function addCourseAndSection(courseId, sectionId, downloadHeader) {
-    return EchoSection.findOrCreate({
-        where: {
-            sectionId: sectionId
-        },
-        defaults: {
-            sectionId: sectionId,
-            courseId: courseId,
-            json: {
-                downloadHeader: downloadHeader
-            }
-        }
-    });
-}
-
-async function addMedia(videoURL, sourceType, siteSpecificJSON) {
-    var media = await Media.findOrCreate({
-        where: {
-            videoURL: videoURL
-        },
-        defaults: {
-            videoURL: videoURL,
-            sourceType: sourceType,
-            siteSpecificJSON: JSON.stringify(siteSpecificJSON)
-        }
-    });
-    return media[0].id;
-}
-
-async function addMSTranscriptionTask(mediaId) {
-    var task = await MSTranscriptionTask.findOrCreate({
-        where: {
-        mediaId: mediaId
-        },
-        defaults: {
-        id: uuid(),
-        mediaId: mediaId
-        }
-    });
-    return task[0].id;
-}
-
-// Return taskId
-async function addToMediaAndMSTranscriptionTask(videoURL, sourceType, siteSpecificJSON, courseOfferingId) {
-    var mediaId = await addMedia(videoURL, sourceType, siteSpecificJSON);
-    await addCourseOfferingMedia(courseOfferingId, mediaId, siteSpecificJSON);
-    var taskId = await addMSTranscriptionTask(mediaId);
-    return taskId;
-}
-
-// function getTask(taskId) {
-//     return MSTranscriptionTask.findById(taskId);
-// }
-//
-// function getMedia(mediaId) {
-//     return Media.findById(mediaId);
-// }
-//
-// function getMediaByTask(taskId) {
-//     return getTask(taskId).then(task => getMedia(task.mediaId));
-// }
-//
-// function getEchoSection(sectionId) {
-//     return EchoSection.findById(sectionId);
-// }
-
-/* findOrCreate University first,
- * then findOrCreate user, and append universityId
- *
- * This is not a suggested way according to the documentation,
- * though I haven't found a better and correct implementation.
- */
-function addUser(user) {
-
-  return University.findOrCreate({
-    where : { universityName : user.university }
-  }).then((result) => {
-    var universityInfo = result[0].dataValues; // result is an array of json object
-    /* create the User or find the User */
-    return User.findOrCreate({
-      where : { mailId : user.mailId },
-      defaults : {
-        id: uuid(),
-        firstName : user.firstName,
-        lastName : user.lastName,
-        password : user.password,
-        passwordToken : '',
-        verified : false,
-        verifiedId : user.verifiedId,
-        universityId : universityInfo.id,
-        googleId : user.googleId,
-      }
-    }).then(result => {
-      return result[0].dataValues;
-    }).catch(err => perror(user, err)); /* User.findOrCreate() */
-  }).catch(err => perror(user, err)); /* University.findOrCreate() */
-}
-
-/* SELECT * FROM User WHERE mailId=email LIMIT 1*/
-// function getUserByEmail(email) {
-//   /* Since the email should be unique,
-//    * findOne() is sufficient
-//    */
-//   return User.findOne({
-//     where : { mailId : email }
-//   }).then(result => {
-//
-//     try {
-//       return result.dataValues;
-//     } catch (err) {
-//       perror(err);
-//       return null;
-//     }
-//
-//   }).catch(err => perror(err));
-// }
-//
-// /* SELECT * FROM User WHERE googleId=profileId LIMIT 1*/
-// function getUserByGoogleId(profileId) {
-//     /* Since the email should be unique,
-//      * findOne() is sufficient
-//      */
-//     return User.findOne({
-//         where: { googleId: profileId }
-//     }).then(result => {
-//         if (result) {
-//             return result.dataValues;
-//         } else {
-//             return null;
-//         }
-//     }).catch(err => perror(err));
-// }
-
-/* UPDATE User SET verifiedId = verifiedId WHERE mailId = email */
-function verifyUser(verifiedId, email) {
-  return User.update( {
-    verified : true,
-  }, {
-    where : {
-      mailId : email,
-      verifiedId : verifiedId,
-    }
-  }).catch(err => perror({verifiedId : verifiedId, mailId : email}, err));
-}
-
-function setUserName(name, email) {
-  return User.update({
-    firstName : name.firstName,
-    lastName : name.lastName,
-  }, {
-    where : {
-      mailId : email,
-    }
-  }).catch(err => perror({mailId : email, firstName : name.firstName, lastName : name.lastName}, err));
-}
-
-function setUserPassword(newPassword, email) {
-  return User.update({
-    password : newPassword,
-    passwordToken : '',
-  }, {
-    where : {
-      mailId : email,
-    }
-  }).catch(err => perror({mailId : email}, err));
-}
-
-/* findOrCreate university to Table Universities,
- * Shouldn't be used since we
- * currently only support UIUC
- */
-// function getUniversityId(universityName) {
-//
-//   return University.findOrCreate({
-//     where: {
-//       universityName : universityName,
-//     }, defaults: {
-//       id: uuid(),
-//     }
-//   }).then(result => {
-//     return result[0].dataValues;
-//   }).catch(err => perror(err));
-// }
-//
-// /* Assuming input is an array */
-// function getCoursesByTerms(term) {
-//   var termId_list = []
-//   var offeringId_list = [];
-//   var courseId_list = [];
-//
-//   return Term.findAll({ // fetch termId
-//     where : {
-//       termName : { [Op.in] : term } // SELECT * FROM Term WHERE termName IN term
-//     }
-//   }).then((result) => { // fetch offeringId
-//     for (let i = 0; i < result.length; i++) {
-//       termId_list[i] = result[i].dataValues.id;
-//     }
-//     return Offering.findAll({
-//       where : {
-//         termId: { [Op.in] : termId_list } // SELECT * FROM Offering WHERE termId == result:  (list)
-//       }
-//     });
-//   }).then((result) => { // fetch courseId
-//     for (let i = 0; i < result.length; i++) {
-//       offeringId_list[i] = result[i].dataValues.id;
-//     }
-//     return CourseOffering.findAll({
-//       where : {
-//         offeringId : { [Op.in] : offeringId_list} // SELECT * FROM CourseOffering WHERE offeringId IN result
-//       }
-//     });
-//   }).then((result) => { // fetch Courses
-//     for (let i = 0; i < result.length; i++) {
-//       courseId_list[i] = result[i].dataValues.courseId;
-//     }
-//     return getCoursesByIds(courseId_list);
-//   });
-// }
-//
-// /* SELECT * FROM Course WHERE id IN courseId */
-// function getCoursesByIds(courseId) {
-//   return Course.findAll({
-//     where : {
-//       id : {[Op.in] : courseId}
-//     }
-//   }).then(values => {
-//     var jsonList = values.map(value => value.dataValues);
-//     return jsonList;
-//   }).catch(err => perror(err));
-// }
-
-function addLecture(courseId, mediaId, date) {
-  return Lecture.findOrCreate({
-    where : {
-      courseId : courseId,
-      mediaId : mediaId,
-    } ,
-    defaults : {
-      id: uuid(),
-      date : date,
-    },
-  }).then(result => {
-    return result[0].dataValues;
-  }).catch(err => perror({user : undefined}, err));
-}
-
-/* findOrCreate the course,
- * then use provided offeringId,
- * update userOffering table
- */
-function addCourseHelper(id) {
-  return CourseOffering.findOrCreate({
-    where : {
-      courseId : id.courseId,
-      offeringId : id.offeringId,
-    },
-  }).then(result => {
-    var courseOfferingInfo = result[0].dataValues;
-
-    return UserOffering.findOrCreate({
-      where : {
-        userId : id.userId,
-        courseOfferingId : courseOfferingInfo.id,
-        roleId : id.roleId,
-      }, defaults: {
-        id: uuid(),
-      }
-    }).then(result => {
-      return result[0].dataValues;
-    }).catch(err => perror(id, err)); /* UserOffering.findOrCreate() */
-  }).catch(err => perror(id, err)); /*  CourseOffering.findOrCreate() */
-}
-
-// function getCourseId(courseInfo) {
-//   var dept = courseInfo.dept;
-//   return Dept.findOne({
-//     where : {
-//       deptName : dept.name,
-//       acronym : dept.acronym,
-//     }
-//   }).then(result => {
-//     var deptInfo = result.dataValues;
-//     return Course.findOrCreate({
-//       where : {
-//         courseName : courseInfo.courseName,
-//         courseNumber : courseInfo.courseNumber,
-//         deptId : deptInfo.id,
-//       },
-//       defaults : {
-//         id: uuid(),
-//         courseDescription : courseInfo.courseDescription,
-//       }
-//     }).then(result => {
-//       return result[0].dataValues;
-//     }).catch(err => perror(err)); /* Course.findOrCreate() */
-//   }).catch(err => perror(err)); /* Dept.findOne() */
-// }
-//
-// /* getRole() findOrCreate a role */
-// function getRoleId(role) {
-//   return Role.findOrCreate({
-//     where : { roleName : role },
-//     defaults : { id: uuid() }
-//   }).then(result => {
-//     return result[0].dataValues;
-//   }).catch(err => perror(err));
-// }
-//
-// /* findOrCreate term, and return termId */
-// function getTermId(term) {
-//   return Term.findOrCreate({
-//     where : { termName : term },
-//     defaults : { id: uuid() }
-//   }).then( result => {
-//     return result[0].dataValues;
-//   }).catch( err => perror(err));
-// }
-//
-// function getDeptId(dept) {
-//   return Dept.findOrCreate({
-//     where : { deptName : dept.name },
-//     defaults : {
-//       id: uuid(),
-//       acronym : dept.acronym
-//     }
-//   }).then(result => {
-//     return result[0].dataValues;
-//   }).catch(err => perror(err));
-// }
-//
-// /* getOfferingId() findOrCreate an offeringId */
-// function getOfferingId(id, sectionName) {
-//   return Offering.findOrCreate({
-//     where : {
-//       termId : id.termId,
-//       deptId : id.deptId,
-//       universityId : id.universityId,
-//       section : sectionName,
-//     }, defaults : {
-//       id: uuid()
-//     }
-//   }).then(result => {
-//     return result[0].dataValues;
-//   }).catch(err => perror(err));
-// }
-
 /*
- * course = {
- *   courseName, courseNumber, courseDescription,
- *   dept, term, section, deptName, acronym
- *  }
- */
-function addCourse(user, course) {
-  /* if user and courseList are not empty */
-  if(user && course) {
-    var id = { universityId : user.universityId };
-    var role_result = getRoleId( 'Instructor' );
-    var user_result = getUserByEmail( user.mailId );
-    return Promise.all([role_result, user_result]).then( values => {
+// // const perror = utils.perror;
+// // const info = utils.info;
+// // const log = utils.log;
+// //
+// // const Op = Sequelize.Op;
+// // const CourseOffering = models.CourseOffering;
+// // const Course = models.Course;
+// // const Dept = models.Dept;
+// // const EchoSection = models.EchoSection;
+// // const Lecture = models.Lecture;
+// // const Media = models.Media;
+// // const MSTranscriptionTask = models.MSTranscriptionTask;
+// // const Offering = models.Offering;
+// // const Role = models.Role;
+// // const Term = models.Term;
+// // const University = models.University;
+// // const User = models.User;
+// // const UserOffering = models.UserOffering;
+// // const YoutubeChannel = models.YoutubeChannel;
+// // const CourseOfferingMedia = models.CourseOfferingMedia;
 
-      var id = {
-        roleId : values[0].id,
-        userId : values[1].id,
-        universityId : values[1].universityId,
-      }
 
-      /* userId is not matched */
-      if (id.userId != user.id) {
-        throw Error('User ID is not matched');
-      }
-
-      /* add course */
-      let term_result = getTermId( course.term );
-      let dept_result = getDeptId( course.dept );
-      return Promise.all([term_result, dept_result]).then(values => {
-
-        id.termId = values[0].id;
-        id.deptId = values[1].id;
-        let course_result = getCourseId(course);
-        let offering_result = getOfferingId(id, course.section);
-        return Promise.all([course_result, offering_result]).then(values => {
-
-          id.courseId = values[0].id;
-          id.offeringId = values[1].id;
-          return addCourseHelper(id);
-        }).catch(err => { perror(user, err) }); /* end of getOfferingId() */
-      }).catch(err => { perror(user, err) }); /* end of Promise.all */
-    }).catch(err => { perror(user, err) }); /* end of Promise.all */
-  }
-  throw Error('User Info or Course List is empty');
-}
+// // function getAllCourses() {
+// //   return Course.findAll({limit: 8})
+// // }
 //
-// /* Get All Terms */
-// function getTerms() {
-//   return Term.findAll().then(values => {
-//     var result = values.map(value => value.dataValues);
-//     return result;
-//   });
-// }
+// // function addCourseOfferingMedia(courseOfferingId, mediaId, description) {
+// //     return CourseOfferingMedia.findOrCreate({
+// //       where: {
+// //         courseOfferingId: courseOfferingId,
+// //         mediaId: mediaId
+// //       },
+// //       defaults: {
+// //           id: uuid(),
+// //           descpJSON: JSON.stringify(description),
+// //           mediaId: mediaId,
+// //           courseOfferingId: courseOfferingId
+// //       }
+// //     })
+// // }
 //
-// /* Get All Courses by University */
-// function getCoursesByUniversityId( universityId ) {
+// // function getPlaylistByCourseOfferingId(courseOfferingId) {
+// //   return sequelize.query(
+// //    'SELECT mst.videoLocalLocation, mst.srtFileLocation, M.siteSpecificJSON \
+// //     FROM MSTranscriptionTasks AS mst \
+// //     INNER JOIN Media as M on mst.mediaId = M.id \
+// //     INNER JOIN CourseOfferingMedia as com on com.mediaId = M.id \
+// //     WHERE com.courseOfferingId = ?',
+// //    { replacements: [ courseOfferingId ], type: sequelize.QueryTypes.SELECT}).catch(err => perror(err)); /* raw query */
+// // }
 //
-//   return sequelize.query(
-//     'SELECT oid.*, cid.*, coid.id\
-//      FROM \
-//      Offerings oid, CourseOfferings coid, Courses cid \
-//      WHERE \
-//      oid.id = coid.offeringId AND cid.id = coid.courseId AND oid.universityId = ?',
-//      { replacements: [ universityId ], type: sequelize.QueryTypes.SELECT })
-//      .then(values => {
-//        return values.map(value => {
+// // function addYoutubeChannelPlaylist(playlistId, channelId) {
+// //     return YoutubeChannel.findOrCreate({
+// //         where: {
+// //             playlistId: playlistId
+// //         },
+// //         defaults: {
+// //             playlistId: playlistId,
+// //             channelId: channelId
+// //         }
+// //     });
+// // }
+// //
+// // function addCourseAndSection(courseId, sectionId, downloadHeader) {
+// //     return EchoSection.findOrCreate({
+// //         where: {
+// //             sectionId: sectionId
+// //         },
+// //         defaults: {
+// //             sectionId: sectionId,
+// //             courseId: courseId,
+// //             json: {
+// //                 downloadHeader: downloadHeader
+// //             }
+// //         }
+// //     });
+// // }
+// //
+// // async function addMedia(videoURL, sourceType, siteSpecificJSON) {
+// //     var media = await Media.findOrCreate({
+// //         where: {
+// //             videoURL: videoURL
+// //         },
+// //         defaults: {
+// //             videoURL: videoURL,
+// //             sourceType: sourceType,
+// //             siteSpecificJSON: JSON.stringify(siteSpecificJSON)
+// //         }
+// //     });
+// //     return media[0].id;
+// // }
+// //
+// // async function addMSTranscriptionTask(mediaId) {
+// //     var task = await MSTranscriptionTask.findOrCreate({
+// //         where: {
+// //         mediaId: mediaId
+// //         },
+// //         defaults: {
+// //         id: uuid(),
+// //         mediaId: mediaId
+// //         }
+// //     });
+// //     return task[0].id;
+// // }
+// //
+// // // Return taskId
+// // async function addToMediaAndMSTranscriptionTask(videoURL, sourceType, siteSpecificJSON, courseOfferingId) {
+// //     var mediaId = await addMedia(videoURL, sourceType, siteSpecificJSON);
+// //     await addCourseOfferingMedia(courseOfferingId, mediaId, siteSpecificJSON);
+// //     var taskId = await addMSTranscriptionTask(mediaId);
+// //     return taskId;
+// // }
 //
-//          /* move value.id to value.courseOfferingId */
-//          value.courseOfferingId = value.id;
-//          delete value.id;
+// // function getTask(taskId) {
+// //     return MSTranscriptionTask.findById(taskId);
+// // }
+// //
+// // function getMedia(mediaId) {
+// //     return Media.findById(mediaId);
+// // }
+// //
+// // function getMediaByTask(taskId) {
+// //     return getTask(taskId).then(task => getMedia(task.mediaId));
+// // }
+// //
+// // function getEchoSection(sectionId) {
+// //     return EchoSection.findById(sectionId);
+// // }
 //
-//          /* remove time attribute */
-//          delete value.createdAt;
-//          delete value.updatedAt;
-//          return value;
-//        })
-//      }).catch(err => perror(err)); /* raw query */
-// }
+// // /* findOrCreate University first,
+// //  * then findOrCreate user, and append universityId
+// //  *
+// //  * This is not a suggested way according to the documentation,
+// //  * though I haven't found a better and correct implementation.
+// //  */
+// // function addUser(user) {
+// //
+// //   return University.findOrCreate({
+// //     where : { universityName : user.university }
+// //   }).then((result) => {
+// //     var universityInfo = result[0].dataValues; // result is an array of json object
+// //     /* create the User or find the User */
+// //     return User.findOrCreate({
+// //       where : { mailId : user.mailId },
+// //       defaults : {
+// //         id: uuid(),
+// //         firstName : user.firstName,
+// //         lastName : user.lastName,
+// //         password : user.password,
+// //         passwordToken : '',
+// //         verified : false,
+// //         verifiedId : user.verifiedId,
+// //         universityId : universityInfo.id,
+// //         googleId : user.googleId,
+// //       }
+// //     }).then(result => {
+// //       return result[0].dataValues;
+// //     }).catch(err => perror(user, err)); /* User.findOrCreate() */
+// //   }).catch(err => perror(user, err)); /* University.findOrCreate() */
+// // }
 //
-// /* Get All Courses by User info */
-// function getCoursesByUserId( uid ) {
+// /* SELECT * FROM User WHERE mailId=email LIMIT 1*/
+// // function getUserByEmail(email) {
+// //   /* Since the email should be unique,
+// //    * findOne() is sufficient
+// //    */
+// //   return User.findOne({
+// //     where : { mailId : email }
+// //   }).then(result => {
+// //
+// //     try {
+// //       return result.dataValues;
+// //     } catch (err) {
+// //       perror(err);
+// //       return null;
+// //     }
+// //
+// //   }).catch(err => perror(err));
+// // }
+// //
+// // /* SELECT * FROM User WHERE googleId=profileId LIMIT 1*/
+// // function getUserByGoogleId(profileId) {
+// //     /* Since the email should be unique,
+// //      * findOne() is sufficient
+// //      */
+// //     return User.findOne({
+// //         where: { googleId: profileId }
+// //     }).then(result => {
+// //         if (result) {
+// //             return result.dataValues;
+// //         } else {
+// //             return null;
+// //         }
+// //     }).catch(err => perror(err));
+// // }
 //
-//   return UserOffering.findAll({
-//     where : { userId : uid },
-//   }).then(values => {
+// // /* UPDATE User SET verifiedId = verifiedId WHERE mailId = email */
+// // function verifyUser(verifiedId, email) {
+// //   return User.update( {
+// //     verified : true,
+// //   }, {
+// //     where : {
+// //       mailId : email,
+// //       verifiedId : verifiedId,
+// //     }
+// //   }).catch(err => perror({verifiedId : verifiedId, mailId : email}, err));
+// // }
+// //
+// // function setUserName(name, email) {
+// //   return User.update({
+// //     firstName : name.firstName,
+// //     lastName : name.lastName,
+// //   }, {
+// //     where : {
+// //       mailId : email,
+// //     }
+// //   }).catch(err => perror({mailId : email, firstName : name.firstName, lastName : name.lastName}, err));
+// // }
+// //
+// // function setUserPassword(newPassword, email) {
+// //   return User.update({
+// //     password : newPassword,
+// //     passwordToken : '',
+// //   }, {
+// //     where : {
+// //       mailId : email,
+// //     }
+// //   }).catch(err => perror({mailId : email}, err));
+// // }
 //
-//     if(values.length === 0) {
-//       return values;
-//     }
+// /* findOrCreate university to Table Universities,
+//  * Shouldn't be used since we
+//  * currently only support UIUC
+//  */
+// // function getUniversityId(universityName) {
+// //
+// //   return University.findOrCreate({
+// //     where: {
+// //       universityName : universityName,
+// //     }, defaults: {
+// //       id: uuid(),
+// //     }
+// //   }).then(result => {
+// //     return result[0].dataValues;
+// //   }).catch(err => perror(err));
+// // }
+// //
+// // /* Assuming input is an array */
+// // function getCoursesByTerms(term) {
+// //   var termId_list = []
+// //   var offeringId_list = [];
+// //   var courseId_list = [];
+// //
+// //   return Term.findAll({ // fetch termId
+// //     where : {
+// //       termName : { [Op.in] : term } // SELECT * FROM Term WHERE termName IN term
+// //     }
+// //   }).then((result) => { // fetch offeringId
+// //     for (let i = 0; i < result.length; i++) {
+// //       termId_list[i] = result[i].dataValues.id;
+// //     }
+// //     return Offering.findAll({
+// //       where : {
+// //         termId: { [Op.in] : termId_list } // SELECT * FROM Offering WHERE termId == result:  (list)
+// //       }
+// //     });
+// //   }).then((result) => { // fetch courseId
+// //     for (let i = 0; i < result.length; i++) {
+// //       offeringId_list[i] = result[i].dataValues.id;
+// //     }
+// //     return CourseOffering.findAll({
+// //       where : {
+// //         offeringId : { [Op.in] : offeringId_list} // SELECT * FROM CourseOffering WHERE offeringId IN result
+// //       }
+// //     });
+// //   }).then((result) => { // fetch Courses
+// //     for (let i = 0; i < result.length; i++) {
+// //       courseId_list[i] = result[i].dataValues.courseId;
+// //     }
+// //     return getCoursesByIds(courseId_list);
+// //   });
+// // }
+// //
+// // /* SELECT * FROM Course WHERE id IN courseId */
+// // function getCoursesByIds(courseId) {
+// //   return Course.findAll({
+// //     where : {
+// //       id : {[Op.in] : courseId}
+// //     }
+// //   }).then(values => {
+// //     var jsonList = values.map(value => value.dataValues);
+// //     return jsonList;
+// //   }).catch(err => perror(err));
+// // }
+// //
+// // function addLecture(courseId, mediaId, date) {
+// //   return Lecture.findOrCreate({
+// //     where : {
+// //       courseId : courseId,
+// //       mediaId : mediaId,
+// //     } ,
+// //     defaults : {
+// //       id: uuid(),
+// //       date : date,
+// //     },
+// //   }).then(result => {
+// //     return result[0].dataValues;
+// //   }).catch(err => perror({user : undefined}, err));
+// // }
+// //
+// // /* findOrCreate the course,
+// //  * then use provided offeringId,
+// //  * update userOffering table
+// //  */
+// // function addCourseHelper(id) {
+// //   return CourseOffering.findOrCreate({
+// //     where : {
+// //       courseId : id.courseId,
+// //       offeringId : id.offeringId,
+// //     },
+// //   }).then(result => {
+// //     var courseOfferingInfo = result[0].dataValues;
+// //
+// //     return UserOffering.findOrCreate({
+// //       where : {
+// //         userId : id.userId,
+// //         courseOfferingId : courseOfferingInfo.id,
+// //         roleId : id.roleId,
+// //       }, defaults: {
+// //         id: uuid(),
+// //       }
+// //     }).then(result => {
+// //       return result[0].dataValues;
+// //     }).catch(err => perror(id, err)); /* UserOffering.findOrCreate() */
+// //   }).catch(err => perror(id, err)); /*  CourseOffering.findOrCreate() */
+// // }
 //
-//     const courseOfferingIds = values.map(value => value.dataValues.courseOfferingId);
+// // function getCourseId(courseInfo) {
+// //   var dept = courseInfo.dept;
+// //   return Dept.findOne({
+// //     where : {
+// //       deptName : dept.name,
+// //       acronym : dept.acronym,
+// //     }
+// //   }).then(result => {
+// //     var deptInfo = result.dataValues;
+// //     return Course.findOrCreate({
+// //       where : {
+// //         courseName : courseInfo.courseName,
+// //         courseNumber : courseInfo.courseNumber,
+// //         deptId : deptInfo.id,
+// //       },
+// //       defaults : {
+// //         id: uuid(),
+// //         courseDescription : courseInfo.courseDescription,
+// //       }
+// //     }).then(result => {
+// //       return result[0].dataValues;
+// //     }).catch(err => perror(err)); /* Course.findOrCreate() */
+// //   }).catch(err => perror(err)); /* Dept.findOne() */
+// // }
+// //
+// // /* getRole() findOrCreate a role */
+// // function getRoleId(role) {
+// //   return Role.findOrCreate({
+// //     where : { roleName : role },
+// //     defaults : { id: uuid() }
+// //   }).then(result => {
+// //     return result[0].dataValues;
+// //   }).catch(err => perror(err));
+// // }
+// //
+// // /* findOrCreate term, and return termId */
+// // function getTermId(term) {
+// //   return Term.findOrCreate({
+// //     where : { termName : term },
+// //     defaults : { id: uuid() }
+// //   }).then( result => {
+// //     return result[0].dataValues;
+// //   }).catch( err => perror(err));
+// // }
+// //
+// // function getDeptId(dept) {
+// //   return Dept.findOrCreate({
+// //     where : { deptName : dept.name },
+// //     defaults : {
+// //       id: uuid(),
+// //       acronym : dept.acronym
+// //     }
+// //   }).then(result => {
+// //     return result[0].dataValues;
+// //   }).catch(err => perror(err));
+// // }
+// //
+// // /* getOfferingId() findOrCreate an offeringId */
+// // function getOfferingId(id, sectionName) {
+// //   return Offering.findOrCreate({
+// //     where : {
+// //       termId : id.termId,
+// //       deptId : id.deptId,
+// //       universityId : id.universityId,
+// //       section : sectionName,
+// //     }, defaults : {
+// //       id: uuid()
+// //     }
+// //   }).then(result => {
+// //     return result[0].dataValues;
+// //   }).catch(err => perror(err));
+// // }
 //
-//     return sequelize.query(
-//       'SELECT oid.*, cid.*, coid.id\
-//       FROM \
-//       Offerings oid, CourseOfferings coid, Courses cid\
-//       WHERE \
-//       oid.id = coid.offeringId AND cid.id = coid.courseId AND coid.id IN (:coids)',
-//       { replacements: { coids : courseOfferingIds }, type: sequelize.QueryTypes.SELECT })
-//       .then(values => {
+// // /*
+// //  * course = {
+// //  *   courseName, courseNumber, courseDescription,
+// //  *   dept, term, section, deptName, acronym
+// //  *  }
+// //  */
+// // function addCourse(user, course) {
+// //   /* if user and courseList are not empty */
+// //   if(user && course) {
+// //     var id = { universityId : user.universityId };
+// //     var role_result = getRoleId( 'Instructor' );
+// //     var user_result = getUserByEmail( user.mailId );
+// //     return Promise.all([role_result, user_result]).then( values => {
+// //
+// //       var id = {
+// //         roleId : values[0].id,
+// //         userId : values[1].id,
+// //         universityId : values[1].universityId,
+// //       }
+// //
+// //       /* userId is not matched */
+// //       if (id.userId != user.id) {
+// //         throw Error('User ID is not matched');
+// //       }
+// //
+// //       /* add course */
+// //       let term_result = getTermId( course.term );
+// //       let dept_result = getDeptId( course.dept );
+// //       return Promise.all([term_result, dept_result]).then(values => {
+// //
+// //         id.termId = values[0].id;
+// //         id.deptId = values[1].id;
+// //         let course_result = getCourseId(course);
+// //         let offering_result = getOfferingId(id, course.section);
+// //         return Promise.all([course_result, offering_result]).then(values => {
+// //
+// //           id.courseId = values[0].id;
+// //           id.offeringId = values[1].id;
+// //           return addCourseHelper(id);
+// //         }).catch(err => { perror(user, err) }); /* end of getOfferingId() */
+// //       }).catch(err => { perror(user, err) }); /* end of Promise.all */
+// //     }).catch(err => { perror(user, err) }); /* end of Promise.all */
+// //   }
+// //   throw Error('User Info or Course List is empty');
+// // }
+// //
+// // /* Get All Terms */
+// // function getTerms() {
+// //   return Term.findAll().then(values => {
+// //     var result = values.map(value => value.dataValues);
+// //     return result;
+// //   });
+// // }
+// //
+// // /* Get All Courses by University */
+// // function getCoursesByUniversityId( universityId ) {
+// //
+// //   return sequelize.query(
+// //     'SELECT oid.*, cid.*, coid.id\
+// //      FROM \
+// //      Offerings oid, CourseOfferings coid, Courses cid \
+// //      WHERE \
+// //      oid.id = coid.offeringId AND cid.id = coid.courseId AND oid.universityId = ?',
+// //      { replacements: [ universityId ], type: sequelize.QueryTypes.SELECT })
+// //      .then(values => {
+// //        return values.map(value => {
+// //
+// //          /* move value.id to value.courseOfferingId */
+// //          value.courseOfferingId = value.id;
+// //          delete value.id;
+// //
+// //          /* remove time attribute */
+// //          delete value.createdAt;
+// //          delete value.updatedAt;
+// //          return value;
+// //        })
+// //      }).catch(err => perror(err)); /* raw query */
+// // }
+// //
+// // /* Get All Courses by User info */
+// // function getCoursesByUserId( uid ) {
+// //
+// //   return UserOffering.findAll({
+// //     where : { userId : uid },
+// //   }).then(values => {
+// //
+// //     if(values.length === 0) {
+// //       return values;
+// //     }
+// //
+// //     const courseOfferingIds = values.map(value => value.dataValues.courseOfferingId);
+// //
+// //     return sequelize.query(
+// //       'SELECT oid.*, cid.*, coid.id\
+// //       FROM \
+// //       Offerings oid, CourseOfferings coid, Courses cid\
+// //       WHERE \
+// //       oid.id = coid.offeringId AND cid.id = coid.courseId AND coid.id IN (:coids)',
+// //       { replacements: { coids : courseOfferingIds }, type: sequelize.QueryTypes.SELECT })
+// //       .then(values => {
+// //
+// //         return values.map(value => {
+// //           /* move value.id to value.courseOfferingId */
+// //           value.courseOfferingId = value.id;
+// //           delete value.id;
+// //
+// //           /* remove time attribute */
+// //           delete value.createdAt;
+// //           delete value.updatedAt;
+// //           return value;
+// //         });
+// //
+// //       }).catch(err => perror(err)); /* rawquery */
+// //     }).catch(err => perror(err)); /* UserOffering.findAll() */
+// // }
 //
-//         return values.map(value => {
-//           /* move value.id to value.courseOfferingId */
-//           value.courseOfferingId = value.id;
-//           delete value.id;
+// // function addStudent(studentId, courseOfferingId) {
+// //   return Role.findOrCreate({
+// //     where : { roleName : 'Student', },
+// //     defaults: { id: uuid() }
+// //   }).then(result => {
+// //     var roleInfo = result[0].dataValues;
+// //
+// //     return UserOffering.findOrCreate({
+// //       where : {
+// //         courseOfferingId : courseOfferingId,
+// //         userId : studentId,
+// //       },
+// //       defaults : {
+// //         id: uuid(),
+// //         roleId : roleInfo.id
+// //       }
+// //     }).then((result, created) => {
+// //       if (created === false) {
+// //         log('YOU HAVE REGISTER ALREADY');
+// //       }
+// //       return result[0].dataValues;
+// //     }).catch(err => perror({userId : studentId, courseOfferingId : courseOfferingId}, err)); /* UserOffering.findOrCreate() */
+// //   }).catch(err => perror({userId : studentId, courseOfferingId : courseOfferingId}, err));/* Role.findOrCreate() */
+// // }
 //
-//           /* remove time attribute */
-//           delete value.createdAt;
-//           delete value.updatedAt;
-//           return value;
-//         });
+// // function removeStudent (studentId, offeringId) {
+// //   /* you need to add students before remove them, assuming the roleId has been created */
+// //   return Role.findOne({
+// //     where : { roleName : 'Student',}
+// //   }).then(result => {
+// //     var roleId = result.dataValues.id;
+// //     return UserOffering.destroy({
+// //       where : {
+// //         roleId : roleId,
+// //         userId : studentId,
+// //         courseOfferingId : courseOfferingId,
+// //       }
+// //     }).then(result => {
+// //       log('Student has been removed.');
+// //     }).catch(err => perror({userId : studentId, courseOfferingId : courseOfferingId}, err)); /* UserOffering.destroy() */
+// //   }).catch(err => perror({userId : studentId, courseOfferingId : courseOfferingId}, err)); /* Role.findOne() */
+// // }
 //
-//       }).catch(err => perror(err)); /* rawquery */
-//     }).catch(err => perror(err)); /* UserOffering.findAll() */
-// }
-
-function addStudent(studentId, courseOfferingId) {
-  return Role.findOrCreate({
-    where : { roleName : 'Student', },
-    defaults: { id: uuid() }
-  }).then(result => {
-    var roleInfo = result[0].dataValues;
-
-    return UserOffering.findOrCreate({
-      where : {
-        courseOfferingId : courseOfferingId,
-        userId : studentId,
-      },
-      defaults : {
-        id: uuid(),
-        roleId : roleInfo.id
-      }
-    }).then((result, created) => {
-      if (created === false) {
-        log('YOU HAVE REGISTER ALREADY');
-      }
-      return result[0].dataValues;
-    }).catch(err => perror({userId : studentId, courseOfferingId : courseOfferingId}, err)); /* UserOffering.findOrCreate() */
-  }).catch(err => perror({userId : studentId, courseOfferingId : courseOfferingId}, err));/* Role.findOrCreate() */
-}
-
-function removeStudent (studentId, offeringId) {
-  /* you need to add students before remove them, assuming the roleId has been created */
-  return Role.findOne({
-    where : { roleName : 'Student',}
-  }).then(result => {
-    var roleId = result.dataValues.id;
-    return UserOffering.destroy({
-      where : {
-        roleId : roleId,
-        userId : studentId,
-        courseOfferingId : courseOfferingId,
-      }
-    }).then(result => {
-      log('Student has been removed.');
-    }).catch(err => perror({userId : studentId, courseOfferingId : courseOfferingId}, err)); /* UserOffering.destroy() */
-  }).catch(err => perror({userId : studentId, courseOfferingId : courseOfferingId}, err)); /* Role.findOne() */
-}
-
-// function getUniversityName (universityId) {
-//   return University.findById(universityId);
-// }
+// // function getUniversityName (universityId) {
+// //   return University.findById(universityId);
+// // }
+// //
+// // function getTermsById (ids) {
+// //   return Term.findAll({
+// //     where : {
+// //       id : { [Op.in] : ids }
+// //     }
+// //   }).then(values => {
+// //     return values.map(value => value.dataValues);
+// //   }).catch(err => perror(err))
+// // }
+// //
+// // function getDeptsById (ids) {
+// //   return Dept.findAll({
+// //     where : {
+// //       id : { [Op.in] : ids }
+// //     }
+// //   }).then(values => {
+// //     return values.map(value => value.dataValues);
+// //   }).catch(err => perror(err))
+// // }
+// //
+// // function getSectionsById (ids) {
+// //   return Offering.findAll({
+// //     where : {
+// //       id : { [Op.in] : ids }
+// //     },
+// //   }).then(values => {
+// //     return values.map(value => value.dataValues);
+// //   }).catch(err => perror(err))
+// // }
+// //
+// // function getInstructorsByCourseOfferingId (ids) {
+// //   /* empty input */
+// //   if (ids.length === 0) {
+// //     return ids;
+// //   }
+// //
+// //   return Role.findOne({
+// //     /* find Instructor's uuid */
+// //     where : { roleName : 'Instructor', },
+// //   }).then(result => {
+// //     var instructorId = result.dataValues.id;
+// //
+// //     /*
+// //      * query to fetch informations based on courseOfferingIds and roleId
+// //      *
+// //      * only select relative information
+// //      */
+// //     return sequelize.query(
+// //       'SELECT uoid.courseOfferingId, uid.id, uid.firstName, uid.lastName, uid.mailId \
+// //       FROM \
+// //       UserOfferings uoid, Users uid \
+// //       WHERE \
+// //       uoid.courseOfferingId IN (:coids) AND uoid.roleId = :rid AND uoid.userId = uid.id',
+// //       { replacements: { coids : ids, rid : instructorId }, type: sequelize.QueryTypes.SELECT })
+// //       .catch(err => perror(err)); /* sequelize.query() */
+// //     }).catch(err => perror(err)); /* Role.findOne() */
+// // }
 //
-// function getTermsById (ids) {
-//   return Term.findAll({
+// function validateUserAccess( courseOfferingId, userId ) {
+//   return UserOffering.findOne({
 //     where : {
-//       id : { [Op.in] : ids }
-//     }
-//   }).then(values => {
-//     return values.map(value => value.dataValues);
-//   }).catch(err => perror(err))
-// }
-//
-// function getDeptsById (ids) {
-//   return Dept.findAll({
-//     where : {
-//       id : { [Op.in] : ids }
-//     }
-//   }).then(values => {
-//     return values.map(value => value.dataValues);
-//   }).catch(err => perror(err))
-// }
-//
-// function getSectionsById (ids) {
-//   return Offering.findAll({
-//     where : {
-//       id : { [Op.in] : ids }
-//     },
-//   }).then(values => {
-//     return values.map(value => value.dataValues);
-//   }).catch(err => perror(err))
-// }
-//
-// function getInstructorsByCourseOfferingId (ids) {
-//   /* empty input */
-//   if (ids.length === 0) {
-//     return ids;
-//   }
-//
-//   return Role.findOne({
-//     /* find Instructor's uuid */
-//     where : { roleName : 'Instructor', },
-//   }).then(result => {
-//     var instructorId = result.dataValues.id;
-//
-//     /*
-//      * query to fetch informations based on courseOfferingIds and roleId
-//      *
-//      * only select relative information
-//      */
-//     return sequelize.query(
-//       'SELECT uoid.courseOfferingId, uid.id, uid.firstName, uid.lastName, uid.mailId \
-//       FROM \
-//       UserOfferings uoid, Users uid \
-//       WHERE \
-//       uoid.courseOfferingId IN (:coids) AND uoid.roleId = :rid AND uoid.userId = uid.id',
-//       { replacements: { coids : ids, rid : instructorId }, type: sequelize.QueryTypes.SELECT })
-//       .catch(err => perror(err)); /* sequelize.query() */
-//     }).catch(err => perror(err)); /* Role.findOne() */
-// }
-
-function validateUserAccess( courseOfferingId, userId ) {
-  return UserOffering.findOne({
-    where : {
-      courseOfferingId : courseOfferingId,
-      userId : userId,
-    }
-  }).then(result => {
-    return result.dataValues;
-  }).catch(err => perror({userId : studentId, courseOfferingId : courseOfferingId}, err));
-}
-
-function getCourseByOfferingId(offeringId) {
-  return CourseOffering.findAll({
-    where : { offeringId : offeringId }
-  }).then(values => {
-    var courseInfos = values.map(value => value.dataValues);
-    var courseIds = courseInfos.map(courseInfo => courseInfo.id);
-    return Course.findAll({
-      where : {
-        id : { [Op.in] : courseIds },
-      }
-    }).then(values => {
-      return values.map(value => value.dataValues);
-    }).catch(err => perror({offeringId : offeringId}, err)); /* Course.findAll() */
-  }).catch(err => perror({offeringId : offeringId}, err)); /* CourseOffering.findAll() */
-}
-
-// function getDept(deptId) {
-//   return Dept.findOne({
-//     where : {
-//       id : deptId,
+//       courseOfferingId : courseOfferingId,
+//       userId : userId,
 //     }
 //   }).then(result => {
 //     return result.dataValues;
-//   }).catch(err => perror(err));
+//   }).catch(err => perror({userId : studentId, courseOfferingId : courseOfferingId}, err));
 // }
-
-function addPasswordToken(userInfo, token) {
-  return User.update({
-    passwordToken : token,
-  }, {
-    where : {
-      id : userInfo.id,
-      mailId : userInfo.mailId,
-    }
-  }).catch(err => perror(userInfo, err));
-}
-
-function setUserRole(userId, role) {
-  return Role.findOrCreate({
-    where : { roleName : role },
-    defaults : { id: uuid() }
-  }).then(result => {
-    const roleInfo = reuslt[0].dataValues;
-    return User.update({
-      roleId : roleInfo.id,
-    }, {
-      where : { id : userId },
-    }).catch(err => perror({userId : userId}, err)); /* User.update() */
-  }).catch(err => perror({userId : userId}, err)); /* Role.findOrCreate() */
-}
+//
+// // function getCourseByOfferingId(offeringId) {
+// //   return CourseOffering.findAll({
+// //     where : { offeringId : offeringId }
+// //   }).then(values => {
+// //     var courseInfos = values.map(value => value.dataValues);
+// //     var courseIds = courseInfos.map(courseInfo => courseInfo.id);
+// //     return Course.findAll({
+// //       where : {
+// //         id : { [Op.in] : courseIds },
+// //       }
+// //     }).then(values => {
+// //       return values.map(value => value.dataValues);
+// //     }).catch(err => perror({offeringId : offeringId}, err)); /* Course.findAll() */
+// //   }).catch(err => perror({offeringId : offeringId}, err)); /* CourseOffering.findAll() */
+// // }
+//
+// // function getDept(deptId) {
+// //   return Dept.findOne({
+// //     where : {
+// //       id : deptId,
+// //     }
+// //   }).then(result => {
+// //     return result.dataValues;
+// //   }).catch(err => perror(err));
+// // }
+//
+// // function addPasswordToken(userInfo, token) {
+// //   return User.update({
+// //     passwordToken : token,
+// //   }, {
+// //     where : {
+// //       id : userInfo.id,
+// //       mailId : userInfo.mailId,
+// //     }
+// //   }).catch(err => perror(userInfo, err));
+// // }
+//
+// // function setUserRole(userId, role) {
+// //   return Role.findOrCreate({
+// //     where : { roleName : role },
+// //     defaults : { id: uuid() }
+// //   }).then(result => {
+// //     const roleInfo = reuslt[0].dataValues;
+// //     return User.update({
+// //       roleId : roleInfo.id,
+// //     }, {
+// //       where : { id : userId },
+// //     }).catch(err => perror({userId : userId}, err)); /* User.update() */
+// //   }).catch(err => perror({userId : userId}, err)); /* Role.findOrCreate() */
+// // }
 
 module.exports = {
     models: models,
+    addCourseOfferingMedia: adder.addCourseOfferingMedia,
+    addCourseAndSection: adder.addCourseAndSection,
+    addMedia: adder.addMedia,
+    addMSTranscriptionTask: adder.addMSTranscriptionTask,
+    addToMediaAndMSTranscriptionTask: adder.addToMediaAndMSTranscriptionTask,
+    addLecture : adder.addLecture,
+    addCourse : adder.addCourse,
+    createUser: adder.addUser,
+    setPasswordToken : updater.setPasswordToken,
+    addStudent : adder.addStudent,
+    removeStudent : remover.removeStudent,
+    verifyUser: updater.verifyUser,
+    setUserName: updater.setUserName,
+    setUserPassword : updater.setUserPassword,
+    validateUserAccess : updater.validateUserAccess,
     getAllCourses: getter.getAllCourses,
-    addCourseOfferingMedia: addCourseOfferingMedia,
     getPlaylistByCourseOfferingId: getter.getPlaylistByCourseOfferingId,
-    addCourseAndSection: addCourseAndSection,
-    addMedia: addMedia,
-    addMSTranscriptionTask: addMSTranscriptionTask,
-    addToMediaAndMSTranscriptionTask: addToMediaAndMSTranscriptionTask,
-    addLecture : addLecture,
-    addCourse : addCourse,
-    addPasswordToken : addPasswordToken,
-    addStudent : addStudent,
-    removeStudent : removeStudent,
-    createUser: addUser,
-    setUserName: setUserName,
-    setUserPassword : setUserPassword,
-    verifyUser: verifyUser,
     getTask: getter.getTask,
     getMedia: getter.getMedia,
     getMediaByTask: getter.getMediaByTask,
@@ -743,5 +748,4 @@ module.exports = {
     getDeptsById : getter.getDeptsById,
     getSectionsById : getter.getSectionsById,
     getInstructorsByCourseOfferingId : getter.getInstructorsByCourseOfferingId,
-    validateUserAccess : validateUserAccess,
 }
