+++ package.json
@@ -11,6 +11,7 @@
     "child-process-promise": "^2.2.1",
     "connect-flash": "^0.1.1",
     "cookie-parser": "^1.4.0",
+    "csv-writer": "1.2.0",
     "data-action": "1.0.1",
     "dotenv": "0.5.1",
     "ejs": "2.5.5",
